requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
exceptiongroup==1.2.0
fastapi==0.109.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
h11==0.14.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
multitasking==0.0.10
numpy==1.22.0
pandas==1.3.5
pandas-datareader==0.10.0
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2021.3
requests==2.27.1
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
tk==0.1.0
typing_extensions==4.9.0
urllib3==1.26.8
uvicorn==0.27.1
Werkzeug==2.0.2
yfinance==0.1.68
